# Conflicts:
#	bevy_srv/salt/common.sls

+++ join-bevy.sh
@@ -7,4 +7,6 @@ if [ ! -d $SCRIPTPATH ]; then
   popd
 fi
 /usr/bin/env python3 $SCRIPTPATH/configure_machine/bootstrap_bevy_member_here.py "$@"
-cp -n $SCRIPTPATH/vgr .
+if [ ! -f vgr ]; then
+  cp -n $SCRIPTPATH/vgr .
+fi
